major refactoring
I hate myself for dumping this massive patch, so please forgive me.

I have been working on this on/off for a few months now, with a big
summer-sized hole in the middle. The code has become too much
intertwined and would be a PITA to have proper segmented patches. As I
am the only contributor, I guess no one will mind ¯\_(ツ)_/¯.

This _massive_ patch mainly reworks the DBusServerBase and uses treelib
to save the server topology and implements the XML introspection.

Signed-off-by: Filipe Laíns <lains@archlinux.org>