dbus_objects: add support for DBus properties
This required a major refactoring of the decorators, we now use object
descriptors.

https://docs.python.org/3/reference/datamodel.html#object.__get__
https://docs.python.org/3/howto/descriptor.html

The current XML generation is incorrect, properties are annotated as
methods. That should be fixed in a later commit.

dbus_objects.signature should be refactored because having DBusSignature
class like we currently do does not make sense with the new storage
model. Most of the glue code there should be moved to the descriptors.
This will also allow us to cleanly fix the wrong XML generation for
properties.

Signed-off-by: Filipe Laíns <lains@riseup.net>